+++ lib/grape-swagger/doc_methods.rb
@@ -10,7 +10,7 @@ require 'grape-swagger/doc_methods/tag_name_description'
 require 'grape-swagger/doc_methods/parse_params'
 require 'grape-swagger/doc_methods/move_params'
 require 'grape-swagger/doc_methods/headers'
-require 'grape-swagger/doc_methods/required_attributes'
+require 'grape-swagger/doc_methods/build_model_definition'
 
 module GrapeSwagger
   module DocMethods
